refactor: switch to automatic binding and clean up the demo
- Replace 11 manual bind() calls with BindingContext.of() and
  @UseProperty dotted paths for the nested Address/EmergencyContact
  fields
- Make the saved-employee accordion single-open (setMultiple(false))
- Rename project webforj-data-bind -> webforj-databinding (folder,
  artifact,
  package com.webforj.databinding) and list it in the root README
- Remove unused EmployeeService, seed data, Address.oneLine(),
  decorative
  JPA annotations and the jakarta.persistence/validation-api
  dependencies
- Rename roleBox -> role so it auto-binds by name; span Role full width
- Pin webforJ 26.01 release, fix stale comments, banner version and
  web.xml
  display-name
