+++ README.md
@@ -15,6 +15,7 @@ webforj-todo | [Source](https://github.com/webforj/built-with-webforj/tree/main/
 webforj-rest | [Source](https://github.com/webforj/built-with-webforj/tree/main/webforj-rest) | `25.03` | A Spring Boot-powered webforJ application demonstrating two approaches for handling paginated REST API data. Features side-by-side comparison of `CollectionRepository` (in-memory pagination) and `DelegatingRepository` (lazy-loading pagination) patterns with a customer management interface. Showcases webforJ's `Table` component with page-based `Navigator` layout, REST API integration, and randomly generated test data. Perfect for learning how to choose between eager and lazy data loading strategies in webforJ applications.
 webforj-ghostai | [Source](https://github.com/webforj/built-with-webforj/tree/main/webforj-ghostai) | `25.11` | An AI chat demo built with webforJ and Spring AI. Features streaming markdown rendering with the new `MarkdownViewer` component, predictive text input, chat memory, and MCP integration with the webforJ MCP Server for real-time documentation queries. Showcases how to build ChatGPT-style interfaces in pure Java using Google Gemini.
 webforj-bookstore | [Source](https://github.com/webforj/built-with-webforj/tree/main/webforj-bookstore) | `25.11` | A book inventory management system built with webforJ and Spring Boot, demonstrating CRUD operations, role-based access control with Spring Security, and data filtering. Features genre management with colored chips, full-text search, sortable tables with custom renderers, and admin-only views. Uses H2 in-memory database with sample data generated via Java Faker.
+webforj-databinding | [Source](https://github.com/webforj/built-with-webforj/tree/main/webforj-databinding) | `26.01` | An employee onboarding form demonstrating webforJ data binding over nested Java beans, packaged as a traditional WAR and run on the Jetty Maven plugin. A single `BindingContext` created with automatic binding (`BindingContext.of`) wires every field, reaching nested `Address` and `EmergencyContact` beans through `@UseProperty` dotted property paths like `address.street`. Jakarta Validation cascades through the nested beans via `@Valid`, and saving opens a dialog presenting the record in a single-open `Accordion`.
 webforj-focustracker | [Source](https://github.com/webforj/built-with-webforj/tree/main/webforj-focustracker) | `25.10` | A focus timer demonstrating Progressive Web App capabilities in webforJ. Features installable app support via `@AppProfile`, desktop notifications with `DesktopNotification.show()`, and dynamic app icon badges using the App Badge API. Shows how Java developers can access modern web platform features traditionally limited to JavaScript.
 [PingPal↗️](https://www.pingpal.dev/) | [Source](https://github.com/webforj/built-with-webforj/tree/main/pingpal) | `25.02` | **🏆 First Place Winner - Built with webforJ Contest** • A Postman-inspired API testing and development tool created by [Rick van Baalen](https://www.linkedin.com/in/rickvanbaalen/). PingPal provides an interface for testing and managing API requests, featuring authentication support, custom headers, formatted responses, and request history. This contest-winning project showcases the power of webforJ for building professional developer tools.
 [startforJ↗️](https://docs.webforj.com/startforj/) | | | The startforJ app generates a minimal starter project based on various available archetypes that includes all required dependencies, configuration files, and a pre-wired layout - ready to build on.
