--- webforj-data-bind/src/main/resources/static/app.css
@@ -1,12 +1,0 @@
-/*
- * Define application theme
- *
- * See the documentation for details:
- * https://docs.webforj.com/docs/styling/overview
- */
-
-/* dwc-window-center is the scroll container — make it actually scroll so the saved-employees
-   table at the bottom of the page becomes reachable on small viewports. */
-dwc-window-center {
-  overflow-y: auto;
-}
