+++ webforj-databinding/src/main/java/com/webforj/databinding/Application.java
@@ -1,4 +1,4 @@
-package com.webforj.databind;
+package com.webforj.databinding;
 
 import com.webforj.App;
 import com.webforj.annotation.AppProfile;
@@ -6,9 +6,9 @@ import com.webforj.annotation.AppTheme;
 import com.webforj.annotation.Routify;
 import com.webforj.annotation.StyleSheet;
 
-@Routify(packages = "com.webforj.databind.views")
+@Routify(packages = "com.webforj.databinding.views")
 @StyleSheet("ws://app.css")
 @AppTheme("system")
-@AppProfile(name = "webforj-databind", shortName = "webforj-databind")
+@AppProfile(name = "webforj-databinding", shortName = "webforj-databinding")
 public class Application extends App {
 }
