+++ webforj-databinding/src/main/resources/banner.txt
@@ -5,4 +5,4 @@ ${AnsiColor.BRIGHT_MAGENTA}██║███╗██║██╔══╝  █
 ${AnsiColor.MAGENTA}╚███╔███╔╝███████╗██████╔╝██║     ╚██████╔╝██║  ██║╚█████╔╝
 ${AnsiColor.BRIGHT_MAGENTA} ╚══╝╚══╝ ╚══════╝╚═════╝ ╚═╝      ╚═════╝ ╚═╝  ╚═╝ ╚════╝
 ${AnsiColor.DEFAULT}
-  :: ${AnsiColor.BRIGHT_CYAN}webforJ${AnsiColor.DEFAULT} (${AnsiColor.BRIGHT_MAGENTA}26.00${AnsiColor.DEFAULT}) ::
+  :: ${AnsiColor.BRIGHT_CYAN}webforJ${AnsiColor.DEFAULT} (${AnsiColor.BRIGHT_MAGENTA}26.01${AnsiColor.DEFAULT}) ::
